pkp/pkp-lib#9678 removed schedule task table and dao
